#629 added do untill X Support modved patch to seperate File
