+++ Source/ProjectRimFactory/Common/HarmonyPatches/UpdateResourceCountsPatch.cs
@@ -7,44 +7,15 @@ using Verse;
 using UnityEngine;
 using HarmonyLib;
 using System.Threading.Tasks;
+using Verse.Noise;
 
 namespace ProjectRimFactory.Common.HarmonyPatches
 {
-    
+
     public interface IAssemblerQueue
     {
         Map Map { get; }
         List<Thing> GetThingQueue();
-    }
-
-
-    //Art & maybe other things too need a seperate patch
-    [HarmonyPatch(typeof(RecipeWorkerCounter), "CountProducts")]
-    class Patch_CountProducts_AssemblerQueue
-    {
-        
-        static void Postfix(RecipeWorkerCounter __instance,ref int __result, Bill_Production bill)
-        {
-            if (bill.includeFromZone == null) {
-                int i = 0;
-                ThingDef targetDef = __instance.recipe.products[0].thingDef;
-                PRFGameComponent gamecomp = Current.Game.GetComponent<PRFGameComponent>();
-
-                for (i = 0; i < gamecomp.AssemblerQueue.Count; i++)
-                {
-                    //Don't count Recorces of other maps
-                    if (bill.Map != gamecomp.AssemblerQueue[i].Map) continue;
-                    foreach (Thing heldThing in gamecomp.AssemblerQueue[i].GetThingQueue())
-                    {
-                        Thing innerIfMinified = heldThing.GetInnerIfMinified();
-                        if (innerIfMinified.def == targetDef)
-                        {
-                            __result += innerIfMinified.stackCount;
-                        }
-                    }
-                }
-            }
-        }
     }
